requirements.txt
flask==3.0.0
groq==0.9.0
httpx==0.27.0
python-dotenv==1.0.0
flask-cors==4.0.0
psycopg[binary]==3.2.1